migrate environment tracking to martin_core
