+++ martin-core/src/config/mod.rs
@@ -1,3 +1,6 @@
 mod cfg_containers;
 
+// Environment variable access with substitution tracking.
+pub mod env;
+
 pub use cfg_containers::*;
