+++ martin/src/args/mod.rs
@@ -1,9 +1,6 @@
 mod connections;
 pub use connections::{Arguments, State};
 
-mod environment;
-pub use environment::{Env, OsEnv};
-
 #[cfg(feature = "postgres")]
 mod pg;
 #[cfg(feature = "postgres")]
