+++ martin/src/bin/martin.rs
@@ -1,8 +1,9 @@
 use clap::Parser;
 use log::{error, info, log_enabled};
-use martin::args::{Args, OsEnv};
+use martin::args::Args;
 use martin::srv::new_server;
 use martin::{Config, MartinResult, read_config};
+use martin_core::config::env::OsEnv;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
